perf: zero-alloc hot path (pooled search scratch + integer timestamp parse)
Hot-path optimizations for the 0.425-CPU / 160MB budget, on top of the v1.0.7
best-first search. Detection is unchanged (bit-identical), only CPU/GC drop.

- Pool the search scratch per thread (ThreadLocal): FrontierHeap, the queryCodes
  IntArray and the KNearest buffer are reused across queries with a reset()
  instead of allocated per request. At budget 32768 the FrontierHeap alone was
  the dominant allocator (growing primitive arrays, ~MBs/s of churn) in a heap
  with only ~23MB free above the 97MB index -> GC pressure on p99. The CIO worker
  count is tiny (1 under -R:ActiveProcessorCount=1) and each query is a
  synchronous, non-reentrant call, so ThreadLocal reuse is safe.

- Parse the fixed UTC timestamp "YYYY-MM-DDTHH:MM:SSZ" with integer arithmetic
  (Sakamoto day-of-week + Hinnant days-from-civil) instead of Instant.parse x2
  per request. Falls back to Instant.parse for any non-fixed shape, so behavior
  is unchanged. A 6048-case sweep test asserts the fast path is BIT-IDENTICAL to
  the Instant-based computation (a 1-bit drift could flip a quantized k-NN code).

- Cleanups: drop the unused squaredDistance(ShortArray,ShortArray) overload
  (the quantized test oracle now uses logical codes like production), remove the
  obsolete offline tools SampleSweep (self-referential metric that caused earlier
  misdiagnoses) and MemProbe, fix stale int8 -> int16 test docstrings.

+++ src/test/kotlin/VectorIndexValidationTest.kt
@@ -14,7 +14,7 @@ import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 
 /**
- * Validates the chosen design (bucketing + VP-Tree int8) against the brute-force
+ * Validates the chosen design (bucketing + VP-Tree int16) against the brute-force
  * oracles on the example dataset (tasks 1.3, 1.4 and 1.6).
  */
 class VectorIndexValidationTest {
@@ -49,7 +49,7 @@ class VectorIndexValidationTest {
     }
 
     @Test
-    fun `int8 quantization preserves the decision versus the double oracle`() {
+    fun `int16 quantization preserves the decision versus the double oracle`() {
         // Quantization may, in principle, flip a borderline neighbor. Confirm it
         // does not change the approve/deny decision on the example set.
         val qs = queries()
